Allow importing of graphql-docs-generator inside other projects
