+++ packages/amplify-graphql-docs-generator/__integration__/e2e.test.ts
@@ -1,6 +1,6 @@
 import { resolve } from 'path'
 import * as fs from 'fs';
-import generate from '../src'
+import generate from '../src/cliFileHandler'
 
 describe('end 2 end tests', () => {
   const schemaPath = resolve(__dirname, '../fixtures/schema.json');
